any ability commit
made a way to give any pokemon an ability with code. had to add new var to the ability info. then ad to make space because of memory size. removed one save slot and reordered memory storage. And i removed parts of contests and pokemon vars such as ribbon bools